chore: fix circular dep
